No shared memory caches for global device memory
The optimized kernels make use of shared memory caches for some pieces
of device global memory.  This is implemented differently for different
polynomial orders and different CUDA block dimensions (possibly to
overcome device limits on thread block sizes and shared memory limits).

The generic kernel does not make use of any shared memory and is
intended to work for all polynomial orders and CUDA block dimensions.

As of this commit, the optimized kernel only shows 1.04 x speedup
compared to the generic kernel for the nek_cuf16 example on an NVIDIA
P100 GPU.